runtests.sh: Check for test executable.
Previously 'cabal test' would produce errors if you tried
to run it without having built the test executable. Thanks
to Markus Rothe for the suggestion.

+++ tests/runtests.sh
@@ -6,15 +6,19 @@
 # Exit status is number of failed tests.
 TESTPROG=../dist/build/testTeXMathML/testTeXMathML
 failures=0
-for t in *.tex; do
-    $TESTPROG <$t >tmp
-    diff ${t%.tex}.xhtml tmp >tmpdiff
-    if [ "$?" -ne "0" ]; then
-        echo "Test ${t%.tex} FAILED (< expected, > actual):"
-        cat tmpdiff
-        let "failures=$failures+1"
-    else
-        echo "Test ${t%.tex} PASSED"
-    fi
-done
+if [ -f $TESTPROG ]; then
+    for t in *.tex; do
+        $TESTPROG <$t >tmp
+        diff ${t%.tex}.xhtml tmp >tmpdiff
+        if [ "$?" -ne "0" ]; then
+            echo "Test ${t%.tex} FAILED (< expected, > actual):"
+            cat tmpdiff
+            let "failures=$failures+1"
+        else
+            echo "Test ${t%.tex} PASSED"
+        fi
+    done
+else
+    echo "Test executable not built. NOT running tests."
+fi
 exit $failures
